Rename Assignment_Mockup_1 to Notes_Mockup_1
Took some time to learn LaTeX so that I can incorporate into real word use cases for school. Seems like a better way to create a thesis of sorts! Enjoy! I'll keep updating my code! Feel free to copy and paste it on overleaf or the IDE of your choice :)